Code and File Clean up + Static Website Decision
Deleted gradle tutorial class and packages and "extra" chromedrivers that are not crucial for this to work.

Pivoting focus from Selenium, dynamic web scraping, and 5e.tools back to scraping the static HTML website "wikidot.5e".

Wikidot is not as "complete" or robust, but it's a lot more simple and straightforward to be used as an early concept for this project