Some fixes and naming security
Implemented logic to check if the spell's name input is correct.

Can optimize this later, but everytime a spell is checked, a full list of spell is compiled made into a list of strings (This can be optimized so that it only has to be made once instead of making a new list everytime it searches for a spell).

It will then check if the spell exists in the spell list. if not, will ask to check the spelling.

In addition, discord embed fields has max character limit of 1024. So, for now, the bot checks if the description is over 1024 - if it is, it will then it will say so and just post the link to the spell.

In addition, capitalization no longer of the spell name input matters

Additionally, fixes to embeds (added message author's username and avatar), and removed unnecessary elements (search as icon)