Clean up and Multiple Events Handler for Bot
The bot can now handle the Ready Event and MessageCreate event concurrently.

getSpellTables was moved to the SpellSearch (but a general "table Search might be implemented in the future so that both spells and items can use it)

A lot of unused lines removed from Spell class

Spell list is now only stored as a static variable within Spell Search. It is automatically created when a spell search is initiated, taking out the need to create a new list whenever a spell is searched.

new EmbedBuilder class to allow less coupling and bloating. Will be used whenever a spell, and soon, an item is created.

Bot commands are now bot-name agnostic. Commands are now always responding based on "!<insert bot username here>"

New bot commands method in simplebot class. Whenever a message is sent, the bot will check if it has the keywords to run a specific command. This now allows different commands to be implemented (such as an Item Search command in the future)

Removed all castinTime, range, components and duration attributes (and its corresponding getters and setters) in the Spell class. Also removed testing printContents() method. Added checks for rituals and concentration (just in case)

Moved "WIKIDOT_URL" from SpellSearch to parent ScraperService class as every class will use that URL (including possible future ItemSearch class)