Scraper Service Content Overhaul
When getting the contents of a page, Scraper Service now automatically keeps <strong> (bold) and <eml> (italic) formats for every word/string.

This is done by also implementing formatParagraphNodes method

Note: It only goes 1 node deep, so it does not keep any word that is both bolded and italized.

This allows for string format to be kept for both Spells and Items.

As a result, metadata names for spells are already formatted. So formatMetaData method is unneeded and deleted.

Also as a result, format for "At Higher Levels." and "Spell List." is still kept.

This may not be best practice, but I hardcoded in a removal for specifically the formated codes (i just think the manual embded versions looked better...)

ItemSearch class duplicated from SpellSearch, to be used as a prototype