Long Spell Description Handling Complete*
The bot can now search up spells without hassle at all, with the ability of sending multiple embeds message at once

* I have decided that it did not actually look that good. I might instead change it to fit Item's similar format - only 1 very large embed.

However, this was a really good learning experience for myself on how to use flux. Might use it to send additional embeds if a spell or item has a table in it.

Almost forgot to rename the discord Token lol