Final Description Section bug fix & take out function
In the bot class, if an entity has a table or spell, it will use the same function to return embeds and corresponding entity embed. Allows better modularity and reduces code duplication.

There was a bug in which the final description section of a really long item or spell was not included. This was fixed